chore: tasks usage reset after deleting flows
