+++ packages/server/api/src/app/ee/flow-run/cloud-flow-run-hooks.ts
@@ -2,6 +2,7 @@ import { ApEdition } from '@activepieces/shared'
 import { FlowRunHooks } from '../../flows/flow-run/flow-run-hooks'
 import { getEdition } from '../../helper/secret-helper'
 import { tasksLimit } from '../project-plan/tasks-limit'
+import { projectUsageService } from '../../project/usage/project-usage-service'
 
 export const platformRunHooks: FlowRunHooks = {
     async onPreStart({ projectId }: { projectId: string }): Promise<void> {
@@ -18,7 +19,7 @@ export const platformRunHooks: FlowRunHooks = {
     }): Promise<void> {
         const edition = getEdition()
         if ([ApEdition.CLOUD, ApEdition.ENTERPRISE].includes(edition)) {
-            await tasksLimit.incrementOrCreateRedisRecord(
+            await projectUsageService.increaseTasks(
                 projectId,
                 tasks,
             )
