+++ packages/server/api/src/app/ee/project-plan/tasks-limit.ts
@@ -4,17 +4,12 @@ import {
     ErrorCode,
     ProjectId,
     ProjectPlan,
-    isNil,
 } from '@activepieces/shared'
 
 import { projectLimitsService } from './project-plan.service'
 import { exceptionHandler } from 'server-shared'
-import { flowRunService } from '../../flows/flow-run/flow-run-service'
 import { getEdition } from '../../helper/secret-helper'
 import { projectUsageService } from '../../project/usage/project-usage-service'
-import { createRedisClient } from '../../database/redis-connection'
-import { Redis } from 'ioredis'
-import { projectService } from '../../project/project-service'
 
 async function limitTasksPerMonth({
     projectPlan,
@@ -46,7 +41,7 @@ async function limit({ projectId }: { projectId: ProjectId }): Promise<void> {
         if (!projectPlan) {
             return
         }
-        const consumedTasks = await incrementOrCreateRedisRecord(projectId, 0)
+        const consumedTasks = await projectUsageService.increaseTasks(projectId, 0)
         await limitTasksPerMonth({
             consumedTasks,
             projectPlan,
@@ -68,38 +63,4 @@ async function limit({ projectId }: { projectId: ProjectId }): Promise<void> {
 
 export const tasksLimit = {
     limit,
-    incrementOrCreateRedisRecord,
-}
-
-const getRedisConnection = (() => {
-    let redis: Redis | null = null
-
-    return (): Redis => {
-        if (!isNil(redis)) {
-            return redis
-        }
-        redis = createRedisClient()
-        return redis
-    }
-})()
-
-
-async function incrementOrCreateRedisRecord(projectId: string, incrementBy: number): Promise<number> {
-    const project = await projectService.getOneOrThrow(projectId)
-    const billingPeriod = projectUsageService.getCurrentingStartPeriod(project.created)
-    const key = `project-usage:${projectId}:${billingPeriod}`
-    const redis = getRedisConnection()
-    const keyExists = await redis.exists(key)
-
-    if (keyExists === 0) {
-        const consumedTasks = await flowRunService.getTasksUsedAfter({
-            projectId,
-            created: billingPeriod,
-        })
-        await redis.set(key, consumedTasks)
-        return incrementBy
-    }
-    else { 
-        return redis.incrby(key, incrementBy)
-    }
 }
